Move solutions to buried package: project_02
